projet_2/src/requirements.txt
pandas>=2.2.0
SQLAlchemy>=2.0.25
psycopg2-binary>=2.9.9
pyarrow>=15.0.0
pyyaml>=6.0.1
python-dateutil>=2.9.0.post0

